pyproject.toml
[project]
name = "chronoveritas"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "fastapi==0.111.0",
    "uvicorn==0.30.0",
    "pydantic==2.7.0",
    "rank-bm25==0.2.2",
    "scipy==1.13.0",
    "openai>=2.7.2",
    "httpx==0.27.0",
    "pyyaml==6.0.1",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"